+++ GA/scripts/calc.global_activity.zsh
@@ -13,6 +13,8 @@ fmri_dir=$root_dir/fMRI_data
 roi_dir=$fmri_dir/roi
 stats_dir=$fmri_dir/stats
 # ============================================================
+fin_dir=$stats_dir/GLM.MO/tsmean/fan_all
+# ============================================================
 mask=fan.roi.GA.all.nii.gz
 # ============================================================
 work_dir=~/Desktop/temp
@@ -25,17 +27,20 @@ foreach nn ($nn_list)
 	foreach gg (GA GB)
 		subj=$gg$nn
 		foreach run (r01 r02 r03 r04 r05 r06)
+			res=$subj.errts.MO.$run.global_activity.1D
+			if [ -f $fin_dir/$res ]; then
+				continue
+			fi
 			data=$subj.errts.MO.$run.nii.gz
 			cp -n $stats_dir/GLM.MO/$nn/$data $work_dir
 
 			cd $work_dir
-			3dmaskave -quiet -mask $mask $data > $subj.errts.MO.$run.global_activity.1D
+			3dmaskave -quiet -mask $mask $data > $res
 			rm $data
 		end
 	end
 end
 # ============================================================
-fin_dir=$stats_dir/GLM.MO/tsmean/fan_all
 if [ ! -d $fin_dir ]; then
 	mkdir -p -m 755 $fin_dir
 fi
